VECTOR CLASS USED; MOVED DRAWING METHODS TO LIFE.CPP
The board is now implemented as a std::vector<std::vector<bool> >
instead of a 2d char array. This implementation should be mroe C++,
an also more flexible thansk to abilities of the vector class.
Methods for graphical drawing and play are now overloaded methods of
Board::play and Board::draw (display is now draw) instead of separate
methods. Their implementation is now located in life.cpp, for proper
compiling with the command line.